remove pdb2crd queue stuff move pdb2crd helper function to dedicated function file
